backend/requirements.txt
Flask==2.3.2
rdflib==6.3.2
flask-cors==4.0.0
pyvis==0.3.1
python-dotenv==1.0.0